refactor: remove unused code and simplify redeeming and transferring tickets logic in Tickets component
